refactor: change logo extension from svg to png
