+++ src/components/ui/header/Header.module.scss
@@ -9,6 +9,11 @@
   font-weight: 300;
 }
 
+.logo {
+  width: 41px;
+  height: 41px;
+}
+
 .nav {
   max-width: 717px;
   width: 100%;
