+++ src/components/ui/info/Info.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Logo from '@/assets/icons/logo1.png';
+import Logo from '@/assets/icons/logo.png';
 import styles from './Info.module.scss';
 
 function Info() {
